% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TMixClust.R
\docType{data}
\name{best_clust_yeast_obj}
\alias{best_clust_yeast_obj}
\title{TMixClust object containing the optimal clustering solution for the
yeast data.}
\format{A \code{TMixClust} object.}
\usage{
best_clust_yeast_obj
}
\value{
optimal clustering solution for the yeast data
}
\description{
This object contains the result of clustering and
stability analysis
corresponding to the clustering solution with the highest likelihood among 10
different runs of clustering on the yeast data with K=4 clusters.
}
\examples{
# Load the optimal clustering solution for the yeast data
# provided with the TMixClust package
data("best_clust_yeast_obj")

# Print the first lines of the yeast clustering object
head(best_clust_yeast_obj)

}
\references{
Golumbeanu M, Desfarges S, Hernandez C, Quadroni M, Rato S,
Mohammadi P, Telenti A, Beerenwinkel N, Ciuffi A. (2017) Dynamics of
Proteo-Transcriptomic Response to HIV-1 Infection.
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@bsse.ethz.ch}
}
\keyword{datasets}
