% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TMixClust.R
\docType{data}
\name{toy_data_df}
\alias{toy_data_df}
\title{Simulated time-series gene expression data}
\format{A \code{data frame} with 91 rows and 6 columns.
The columns correspond to different time points, while the rownames of the
data frame correspond to gene names.}
\usage{
toy_data_df
}
\value{
toy data
}
\description{
This data set contains a toy example of time-series gene
expression data.
}
\examples{
# Load the toy time series data provided with the TMixClust package
data("toy_data_df")

# Print the first lines of the toy data frame
head(toy_data_df)

}
\references{
Golumbeanu M, Desfarges S, Hernandez C, Quadroni M, Rato S,
Mohammadi P, Telenti A, Beerenwinkel N, Ciuffi A. (2017) Dynamics of
Proteo-Transcriptomic Response to HIV-1 Infection.
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@bsse.ethz.ch}
}
\keyword{datasets}
